Trim the demo to one document and give tw-official a wording section
The demo shipped eight sample documents. Two of them reproduced a real
government letter verbatim, so readers judged the project by wording it
had never authored. Keep getting-started as the single entry point and
drop the rest; the theme gallery is unaffected because every theme's
.demo.tsx is self-contained.

tw-official only ever specified layout — typeface, margins, the hanging
indent — and said nothing about wording, which is where an agent-drafted
公文 actually fails. Add 公文用語 covering: 主旨 as one sentence with a
single 期望語, 依據 as the opening 說明 item, 附件 pointing at a 說明
item rather than listing filenames, a 指代語 table, and a table of
common mistakes. Also correct the typeface note — TW-Kai is the NDC
全字庫 face, not a MOE one.

The theme's own demo broke two of those rules (stacked 期望語, repeated
full name where 旨開 belongs). Fixed, since it is now the only surviving
example of the theme.
