+++ apps/demo/themes/tw-official.md
@@ -30,7 +30,8 @@ mode: light
 ## Typography
 
 - 全文使用**標楷體**。CSS stack：`'"DFKai-SB", "BiauKai", "標楷體", "TW-Kai", "Kaiti TC", serif'`
-  - Windows 為 DFKai-SB、macOS 為 BiauKai；教育部標楷體（TW-Kai）字距最接近公文範本。
+  - Windows 為 DFKai-SB、macOS 為 BiauKai；**全字庫正楷體**（TW-Kai，國發會全字庫免費授權）字距最接近公文範本，
+    且涵蓋公文常見的罕用字與異體字，機關內部統一安裝這套最保險。
 - 機關名稱＋文別：約 20pt（27 px）
 - 本文（主旨／說明／辦法、受文者、正副本、署名）：約 16pt（21 px）
 - 檔號欄、聯絡資訊、頁碼：約 12pt（16 px）
@@ -221,6 +222,69 @@ const Footer = () => {
   );
 };
 ```
+
+## 公文用語
+
+**版面對了、文字不對，公文一樣會被退。** 這一節比上面的版面規範更常出錯，而且是 AI 代擬公文最容易失手的地方——模型會寫出「格式正確但讀起來不像公文」的句子。
+
+### 主旨
+
+- **一句話、一件事**，以 50 字為度。超過就把細節移到說明。
+- 結構固定為 `動詞 ＋ 事項 ＋ 期望語`，例：「檢送本府○○要點修正草案 1 份，請查照。」
+- **期望語只用一個。**「請查照並依說明事項辦理」是疊用，擇一即可。
+- 不分項、不換行、結尾用句號。
+- 背景、緣由、原因一律移到說明，不寫進主旨。
+
+```
+❌ 邇來本府偶有獎狀、聘書、感謝狀、證書等文書出現格式錯誤之情形，
+   檢附用印格式供各局處參考，請查照並轉知所屬。      ← 緣由塞進主旨、期望語疊用
+✅ 檢送本府常用文書用印格式 1 份，請查照並轉知所屬。   ← 一句話講完，緣由留給說明
+```
+
+### 說明
+
+- 第一項慣例是**依據**，交代法令、來文或會議決議：「依據本府 115 年 7 月 20 日○字第 1150000000 號函辦理。」
+- 沒有依據可寫時，第一項直接敘明事實或緣由。
+- 主旨提過的事項，說明裡用**指代語**帶過，不要重複全稱：
+
+  | 指代語     | 指涉對象                               |
+  | ---------- | -------------------------------------- |
+  | 旨揭、旨開 | 主旨提到的人事物（旨揭文書、旨開要點） |
+  | 上開、前開 | 前面已敘明的項目                       |
+  | 該         | 最近提到的對象（該局、該案）           |
+  | 本案       | 本件公文所辦事項                       |
+
+  一份文裡把「獎狀、聘書、感謝狀、證書」寫滿四次，是承辦人最常被退的毛病。
+
+### 附件
+
+- 附件欄寫**指向本文的位置**，不是列檔名：`附件：如說明二`。
+- 本文提到時寫「（如附件）」「詳如附件一」。
+- 附件欄列出的東西，本文一定要有對應敘述；**本文沒提到的附件不要出現在附件欄**。
+
+### 慣用語彙
+
+| 類別       | 用語                                       |
+| ---------- | ------------------------------------------ |
+| 期望語（對下級） | 請查照、請照辦、請辦理見復、請轉知所屬 |
+| 期望語（對上級） | 請鑒核、請核示、請鑒察                 |
+| 期望語（平行）   | 請查照、請惠允、請查照見復             |
+| 准駁語     | 應予照准、未便照准、礙難照辦               |
+| 引敘語     | 依據、奉、准、有關、關於                   |
+| 結束語     | 復如說明、如說明二、無                     |
+
+### 常見錯誤
+
+| 錯誤             | 應改為                                                     |
+| ---------------- | ---------------------------------------------------------- |
+| 是類             | 非公文用語。寫「該類」「這類」，或直接省略                 |
+| 主旨分項或換行   | 主旨必須一段寫完                                           |
+| 期望語疊用       | 「請查照並依說明事項辦理」→ 擇一                           |
+| 重複全稱         | 第二次提到就改用旨揭／該／上開                             |
+| 辦法段重述說明   | 說明交代過的不要再寫一次；沒有新指示就**不要有辦法段**     |
+| 口語與贅字       | 「進行檢討」→「檢討」；「做一個說明」→「說明」             |
+
+> **AI 代擬的公文一定要由承辦人依本機關慣例覆核後才發文。** 各機關的用語習慣、期望語搭配、分項慣例都有差異，這一節寫的是通則，不是任何一個機關的定本。
 
 ## Rules
